refactor(tasks): decompose TasksService god-service into routing/planning/actions services + serializer (#837)
tasks.service.ts was a 1,351-line @Injectable mixing CRUD, agent lifecycle,
review/output actions, AI routing classification, planning/follow-up
orchestration, and a hand-rolled realtime serialization layer (the last
violating the repo rule that serializers live in packages/serializers).

Split into focused units, all behavior-preserving:
- TaskRoutingService — pure routing-decision engine (buildRoutingDecision,
  intent, skill-driven + fallback). inferOutputType is now a [regex,outputType]
  table and the fallback is a per-outputType config map (no deep ternaries).
- TaskActionsService — review-gate transitions (approve/requestChanges/dismiss),
  output-array mutations (keep/unkeep/trash/attach), recordTaskEvent, and the
  realtime broadcast. Collapses the four copy-pasted output methods into one
  mutateOutputIdArray helper and the review methods into one applyReviewTransition.
- TaskPlanningService — planning-thread + follow-up-task orchestration.
- packages/serializers workspace-task-realtime.builder — the realtime snapshot
  (buildWorkspaceTaskRealtimeSnapshot + progress/event/date helpers), replacing
  the in-service serialize* methods. Output is byte-compatible with the previous
  wire shape (frontend still consumes payload.task as a flat object).

TasksService keeps only persistence/CRUD/lifecycle plus thin delegators to the
extracted services, so the controller (and its contract) is unchanged. Renamed
requireAiTask -> requireTask (now public, reused by the extracted services) and
patchAiTaskAndReturn -> patchTaskAndReturn.

Every file in collections/tasks/services is now <500 lines (was 1,351). Adds
task-routing.service.spec and task-planning.service.spec covering the
previously-untestable routing-decision and follow-up-extraction logic.

Closes #837
