+++ packages/serializers/src/server/content/index.ts
@@ -36,3 +36,4 @@ export * from '@serializers/server/content/speech-transcription.serializer';
 export * from '@serializers/server/content/template.serializer';
 export * from '@serializers/server/content/transcript.serializer';
 export * from '@serializers/server/content/workspace-task.serializer';
+export * from '@serializers/server/content/workspace-task-realtime.builder';
